New ODST Armor rebuilt to new aax
The incoming odst armor has been updated to be compliant with the new aax style. It should be able to simply merge on top of the aax pull request assuming that's 100% fine. If not this will require an update. I also made the new odst segments code standard compliant.
