+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/vehicles/armour_standard/config.hpp
@@ -1,0 +1,1 @@
+#include "odst\config.hpp"
