+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/weapons/armour_standard/marines2/config.hpp
@@ -1,0 +1,1170 @@
+// MA Marine Uniforms
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "urbanRed";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "urbanRed";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "urbanRed";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_NS
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "shoulderless";
+	camo = "urbanRed";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "urbanRed";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Medium_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "urbanRed";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Heavy_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "urbanRed";
+	collar = "down";
+};class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "urbanBlue";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "urbanBlue";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "urbanBlue";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_NS
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "shoulderless";
+	camo = "urbanBlue";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "urbanBlue";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "urbanBlue";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "urbanBlue";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "arctic";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "arctic";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "arctic";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_NS
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "shoulderless";
+	camo = "arctic";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "arctic";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Medium_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "arctic";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Heavy_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "arctic";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Light
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "arid";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Medium
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "arid";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Heavy
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "arid";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_NS
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "shoulderless";
+	camo = "arid";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Light_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "arid";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Medium_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "arid";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Arid_Heavy_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "arid";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "tropic";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "tropic";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "tropic";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_NS
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "shoulderless";
+	camo = "tropic";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Light_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "tropic";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Medium_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "tropic";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Tropic_Heavy_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "tropic";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "woodland";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "woodland";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "woodland";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_NS
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "shoulderless";
+	camo = "woodland";
+	collar = "up";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Light_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "light";
+	camo = "woodland";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Medium_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "medium";
+	camo = "woodland";
+	collar = "down";
+};
+class 34thPRC_ArmourStandard_Marines2_Uniform_Woodland_Heavy_NoCollar
+{
+	model = "34thPRC_ArmourStandard_Marines2_Uniform_Urban_Light";
+	variant = "heavy";
+	camo = "woodland";
+	collar = "down";
+};
+
+
+
+// MA CH252 Helmets
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "standard";
+	medic = "false";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "green"
+
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "orange"
+};
+
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "standard";
+	medic = "false";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "standard";
+	medic = "false";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "standard";
+	medic = "false";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "standard";
+	medic = "false";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "false";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "standard";
+	medic = "true";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Urban_Medic_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "urban";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "standard";
+	medic = "true";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arctic_Medic_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arctic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "standard";
+	medic = "true";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Arid_Medic_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "arid";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "standard";
+	medic = "true";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Tropic_Medic_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "tropic";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "standard";
+	medic = "true";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Blue
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "blue";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Orange
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "orange";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Green
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "green";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Purple
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "purple";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_EyePiece_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "eyepiece";
+	medic = "true";
+	eyewear_color = "red";
+};
+class 34thPRC_ArmourStandard_Marines2_CH252_Woodland_Medic_VisorGlasses_Red
+{
+	model = "34thPRC_ArmourStandard_Marines2_CH252_Urban";
+	camo = "woodland";
+	eyewear = "visor";
+	medic = "true";
+	eyewear_color = "red";
+};
