+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_infos/weapons/armour_standard/odst2/config.hpp
@@ -1,0 +1,303 @@
+//MA ODST Vest New
+class 34thPRC_ArmourStandard_ODST_M52_Urban
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="urban";
+	variant="a";
+};
+class 34thPRC_ArmourStandard_ODST_M52_B_Urban
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="urban";
+	variant="b";
+};
+class 34thPRC_ArmourStandard_ODST_M52_C_Urban
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="urban";
+	variant="c";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Marksman_Urban
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="urban";
+	variant="marksman";
+};
+class 34thPRC_ArmourStandard_ODST_M52_CQB_Urban
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="urban";
+	variant="cqb";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Medic_Urban
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="urban";
+	variant="medic";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arid";
+	variant="a";
+};
+class 34thPRC_ArmourStandard_ODST_M52_B_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arid";
+	variant="b";
+};
+class 34thPRC_ArmourStandard_ODST_M52_C_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arid";
+	variant="c";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Marksman_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arid";
+	variant="marksman";
+};
+class 34thPRC_ArmourStandard_ODST_M52_CQB_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arid";
+	variant="cqb";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Medic_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arid";
+	variant="medic";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arctic";
+	variant="a";
+};
+class 34thPRC_ArmourStandard_ODST_M52_B_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arctic";
+	variant="b";
+};
+class 34thPRC_ArmourStandard_ODST_M52_C_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arctic";
+	variant="c";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Marksman_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arctic";
+	variant="marksman";
+};
+class 34thPRC_ArmourStandard_ODST_M52_CQB_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arctic";
+	variant="cqb";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Medic_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="arctic";
+	variant="medic";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="tropic";
+	variant="a";
+};
+class 34thPRC_ArmourStandard_ODST_M52_B_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="tropic";
+	variant="b";
+};
+class 34thPRC_ArmourStandard_ODST_M52_C_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="tropic";
+	variant="c";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Marksman_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="tropic";
+	variant="marksman";
+};
+class 34thPRC_ArmourStandard_ODST_M52_CQB_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="tropic";
+	variant="cqb";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Medic_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="tropic";
+	variant="medic";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="woodland";
+	variant="a";
+};
+class 34thPRC_ArmourStandard_ODST_M52_B_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="woodland";
+	variant="b";
+};
+class 34thPRC_ArmourStandard_ODST_M52_C_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="woodland";
+	variant="c";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Marksman_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="woodland";
+	variant="marksman";
+};
+class 34thPRC_ArmourStandard_ODST_M52_CQB_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="woodland";
+	variant="cqb";
+};
+class 34thPRC_ArmourStandard_ODST_M52_Medic_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_M52_Urban";
+	camo="woodland";
+	variant="medic";
+};
+
+//MA ODST Helmets CH252D New
+class 34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban
+{
+	model= "34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo= "urban";
+	visor= "silver";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_RedVisor_Urban
+{
+	model= "34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo= "urban";
+	visor= "red";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_GoldVisor_Urban
+{
+	model= "34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo= "urban";
+	visor= "gold";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_DP_Urban
+{
+	model= "34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo= "urban";
+	visor= "dp";
+};
+class 34thPRC_ArmourStandard_ODST_CHCH252D_SilverVisor_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="arid";
+	visor="silver";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_RedVisor_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="arid";
+	visor="red";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_GoldVisor_Arid
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="arid";
+	visor="gold";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_DP_Arid
+{
+	model= "34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo= "arid";
+	visor= "dp";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="arctic";
+	visor="silver";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_RedVisor_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="arctic";
+	visor="red";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_GoldVisor_Arctic
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="arctic";
+	visor="gold";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_DP_Arctic
+{
+	model= "34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo= "arctic";
+	visor= "dp";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="tropic";
+	visor="silver";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_RedVisor_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="tropic";
+	visor="red";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_GoldVisor_Tropic
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="tropic";
+	visor="gold";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_DP_Tropic
+{
+	model= "34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo= "tropic";
+	visor= "dp";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="woodland";
+	visor="silver";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_RedVisor_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="woodland";
+	visor="red";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_GoldVisor_Woodland
+{
+	model="34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo="woodland";
+	visor="gold";
+};
+class 34thPRC_ArmourStandard_ODST_CH252D_DP_Woodland
+{
+	model= "34thPRC_ArmourStandard_ODST_CH252D_SilverVisor_Urban";
+	camo= "woodland";
+	visor= "dp";
+};
