+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/config.hpp
@@ -1,0 +1,10 @@
+class cfgWeapons
+{
+	//Include weapons
+	#include "weapons\config.hpp"
+};
+class cfgVehicles
+{
+	//Include vehicles
+	#include "vehicles\config.hpp"
+};
