--- src/Addons/34thPRC_ArmourStandard/data/odst2/34th_ODST_Vest_NOHQ.paa
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:0ab0a499f43ee725f2186da995054cf7d118893224ef6cc0a64eec215388185c
-size 673667
